@@ -1,6 +1,6 @@
 {
   "name": "solidarity",
-  "version": "2.3.1",
+  "version": "3.0.0",
   "description": "Make sure all React Native dependencies are uniform across machines",
   "repository": "https://github.com/infinitered/solidarity",
   "bin": {
@@ -24,7 +24,7 @@
     "snapupdate": "jest -u",
     "welcome": "yarn tsc && npm link",
     "serve:docs": "docsify serve docs",
-    "maintenence": "tryupdate && typesync",
+    "maintenence": "yarn tryupdate && typesync",
     "addTypes": "typesync",
     "tryupdate": "updtr",
     "copyTemplates": "cp -R ./src/templates ./dist/",
@@ -37,47 +37,46 @@
   "license": "MIT",
   "dependencies": {
     "callsite": "^1.0.0",
-    "envinfo": "6.0.1",
-    "fsevents": "^1.2.4",
-    "gluegun": "2.0.7",
-    "json5": "2.1.0",
+    "envinfo": "7.5.0",
+    "gluegun": "4.2.0",
+    "json5": "2.1.1",
     "listr": "^0.14.3",
     "minimist": "^1.2.0",
-    "ramda": "0.26.1",
+    "ramda": "0.27.0",
     "resolve-dir": "^1.0.1"
   },
   "devDependencies": {
     "@types/callsite": "^1.0.30",
-    "@types/cli-table2": "0.2.2",
-    "@types/execa": "^0.9.0",
-    "@types/jasmine": "3.3.5",
-    "@types/jest": "23.3.12",
+    "@types/execa": "2.0.0",
+    "@types/jasmine": "3.5.8",
+    "@types/jest": "25.1.3",
     "@types/json5": "0.0.30",
+    "@types/listr": "^0.14.2",
     "@types/minimist": "^1.2.0",
-    "@types/node": "10.12.18",
-    "@types/prettier": "1.15.2",
-    "@types/ramda": "0.25.46",
-    "@types/tempy": "0.2.0",
-    "ajv": "6.6.2",
-    "all-contributors-cli": "5.6.1",
-    "babel-eslint": "10.0.1",
+    "@types/node": "12.6.6",
+    "@types/prettier": "1.19.0",
+    "@types/ramda": "0.26.43",
+    "@types/tempy": "0.3.0",
+    "ajv": "6.12.0",
+    "all-contributors-cli": "6.14.0",
+    "babel-eslint": "10.1.0",
     "codecov.io": "^0.1.6",
-    "coveralls": "^3.0.2",
-    "danger": "7.0.2",
-    "danger-plugin-spellcheck": "^1.2.3",
-    "docsify-cli": "4.3.0",
+    "coveralls": "3.0.9",
+    "danger": "9.2.10",
+    "danger-plugin-spellcheck": "1.5.0",
+    "docsify-cli": "4.4.0",
     "execa": "1.0.0",
     "jest": "22.4.2",
-    "madge": "3.3.0",
-    "np": "3.1.0",
-    "prettier": "1.15.3",
-    "tempy": "^0.2.1",
+    "madge": "3.7.0",
+    "np": "6.2.0",
+    "prettier": "1.19.1",
+    "tempy": "0.4.0",
     "ts-jest": "22.4.6",
-    "ts-node": "^7.0.1",
-    "tslint": "5.12.0",
-    "tslint-config-standard": "7.1.0",
-    "typescript": "3.2.2",
-    "typesync": "0.4.1",
+    "ts-node": "8.6.2",
+    "tslint": "6.0.0",
+    "tslint-config-standard": "9.0.0",
+    "typescript": "3.8.3",
+    "typesync": "0.6.1",
     "updtr": "^3.1.0"
   },
   "prettier": {
