@@ -45,18 +45,18 @@ describe('without a .solidarity file', () => {
     process.chdir(origCwd)
   })
 
-  // it('should prompt the user', async () => {
-  //   await snapshotCommand.run(context)
-  //   expect(context.prompt.ask.mock.calls).toEqual([
-  //     [
-  //       {
-  //         message: 'No `.solidarity` file found for this project.  Would you like to create one?',
-  //         name: 'createFile',
-  //         type: 'confirm',
-  //       },
-  //     ],
-  //   ])
-  // })
+  it('should prompt the user', async () => {
+    await snapshotCommand.run(context)
+    expect(context.prompt.ask.mock.calls).toEqual([
+      [
+        {
+          message: 'No `.solidarity` file found for this project.  Would you like to create one?',
+          name: 'createFile',
+          type: 'confirm',
+        },
+      ],
+    ])
+  })
 })
 
 describe('with a .solidarity file', () => {
