@@ -15,7 +15,7 @@ module.exports = async (rule: CLIRule, context: SolidarityRunContext): Promise<s
     }
   } else {
     // We try the following in this order
-    // -v, --version, -version
+    // -v, --version, -version, version
     try {
       versionOutput = await system.run(`${rule.binary} -v`)
     } catch (_e) {
@@ -25,7 +25,11 @@ module.exports = async (rule: CLIRule, context: SolidarityRunContext): Promise<s
         try {
           versionOutput = await system.run(`${rule.binary} -version`)
         } catch (_e) {
-          throw 'No version identifier flag for this binary was found'
+          try {
+            versionOutput = await system.run(`${rule.binary} version`)
+          } catch (_e) {
+            throw 'No version identifier flag for this binary was found'
+          }
         }
       }
     }
