@@ -1,7 +1,5 @@
-import { SolidarityRunContext } from '../../../../types'
+import { SolidarityRunContext, SolidarityRule } from '../../../../types'
 
-export default async (context: SolidarityRunContext): Promise<void> => {
-  const { print } = context
-  print.info(`kickoff CLI function`)
-  // return { rule: 'cli' }
+export default async (context: SolidarityRunContext): Promise<SolidarityRule> => {
+  return { rule: 'cli', binary: '' }
 }
