@@ -11,10 +11,9 @@ export default async (context: SolidarityRunContext): Promise<string> => {
   `)
   const userAnswer = await prompt.ask({
     name: 'addRule',
-    type: 'radio',
+    type: 'select',
     message: 'What kind of rule would you like to add?',
-    default: 'cli',
-    choices: ['cli', 'env', 'file', 'dir', 'shell']
+    choices: ['cli', 'env', 'file', 'dir', 'shell'],
   })
 
   return userAnswer.addRule
